Server input receiver
This code acts as an inbetween for the client and the robot, as in order to send the inputs over long distances, the devices must be connected on the same network which isnt possible with the robot (without purchasing more equipment). So all this code does is receive an input from a long distance client over a WAN network, and relays that input to the robot which is on the same local network.